[major] breakdown the huge quantum_manager.py file into a module consiting of several files. Update some imports
