--- sequence/kernel/quantum_manager.py
@@ -1,944 +1,0 @@
-"""This module defines the quantum manager class, to track quantum states.
-
-The states may currently be defined in two possible ways:
-    - KetState
-    - DensityMatrix
-    - FockDensityMatrix
-    - Bell Diagonal
-
-The manager defines an API for interacting with quantum states.
-"""
-from abc import ABC, abstractmethod
-from numpy.typing import NDArray
-from threading import Lock
-from typing import TYPE_CHECKING, Any
-
-if TYPE_CHECKING:
-    from ..components.circuit import Circuit
-    from .quantum_state import State
-
-from qutip_qip.circuit import QubitCircuit
-from qutip_qip.operations import gate_sequence_product, Gate
-from numpy import cumsum, base_repr
-from scipy.sparse import csr_matrix
-from scipy.special import binom
-
-from .quantum_state import KetState, DensityState, BellDiagonalState
-from .quantum_utils import *
-from ..constants import KET_VECTOR_FORMALISM, DENSITY_MATRIX_FORMALISM, FOCK_DENSITY_MATRIX_FORMALISM, BELL_DIAGONAL_STATE_FORMALISM
-
-
-class QuantumManager(ABC):
-    """Class to track and manage quantum states (abstract).
-
-    All states stored are of a single formalism (by default as a ket vector).
-
-    Class Attributes:
-        _registry (dict): mapping of formalism names to manager classes.
-        _global_formalism_lock (Lock): lock for managing global formalism.
-        _global_formalism (str): global formalism.
-
-    Attributes:
-        states (dict[int, State]): mapping of state keys to quantum state objects.
-        _least_available (int): tracking the total number of quantum states in the quantum network
-        truncation (int): maximally allowed number of excited states for elementary subsystems. Default is 1 for qubit.
-        dim (int): subsystem Hilbert space dimension. dim = truncation + 1
-    """
-    _registry: dict = {}
-    _global_formalism_lock = Lock()
-    _global_formalism: str = KET_VECTOR_FORMALISM
-
-    def __init__(self, truncation: int = 1):
-        self.states: dict[int, "State"] = {}
-        self._least_available: int = 0
-        self.truncation = truncation
-        self.dim = self.truncation + 1
-
-    @classmethod
-    def set_global_manager_formalism(cls, formalism: str):
-        """Set the global manager formalism.
-
-        Args:
-            formalism (str): The formalism to set as the global manager formalism.
-        """
-        with cls._global_formalism_lock:
-            if formalism not in cls._registry:
-                raise ValueError(f"Quantum manager '{formalism}' is not registered.")
-            cls._global_formalism = formalism
-
-    @classmethod
-    def get_active_formalism(cls):
-        with cls._global_formalism_lock:
-            return cls._global_formalism
-
-    @classmethod
-    def clear_active_formalism(cls):
-        with cls._global_formalism_lock:
-            cls._global_formalism = KET_VECTOR_FORMALISM
-
-    @classmethod
-    def register(cls, name: str, manager_class=None):
-        """Register a quantum manager class.
-
-        Args:
-            name (str): The name of the quantum manager.
-            manager_class (type, optional): The manager class to register.
-        """
-        if manager_class is not None:
-            cls._registry[name] = manager_class
-            return None
-
-        def decorator(manager_cls):
-            cls._registry[name] = manager_cls
-            return manager_cls
-
-        return decorator
-
-    @classmethod
-    def create(cls, *args, **kwargs) -> 'QuantumManager':
-        """Create a new instance of the quantum manager.
-        """
-        active_formalism = cls.get_active_formalism()
-        if active_formalism not in cls._registry:
-            raise ValueError(f"Quantum manager '{active_formalism}' is not registered.")
-
-        return cls._registry[active_formalism](*args, **kwargs)
-
-    @abstractmethod
-    def new(self, state) -> int:
-        """Method to create a new quantum state.
-
-        Args:
-            state: complex amplitudes of new state. Type depends on type of subclass.
-
-        Returns:
-            int: key for new state generated.
-        """
-        pass
-
-    def get(self, key: int) -> "State":
-        """Method to get quantum state stored at an index.
-
-        Args:
-            key (int): key for quantum state.
-
-        Returns:
-            State: quantum state at supplied key.
-        """
-        return self.states[key]
-
-    @abstractmethod
-    def run_circuit(self, circuit: "Circuit", keys: list[int], meas_samp=None):
-        """Method to run a circuit on a given set of quantum states.
-
-        Args:
-            circuit (Circuit): quantum circuit to apply.
-            keys (list[int]): list of keys for quantum states to apply circuit to.
-            meas_samp (float): random sample used for measurement.
-
-        Returns:
-            dict[int, int]: dictionary mapping qstate keys to measurement results.
-        """
-
-        assert len(keys) == circuit.size, "mismatch between circuit size and supplied qubits"
-        if len(circuit.measured_qubits) > 0:
-            assert meas_samp, "must specify random sample when measuring qubits"
-
-    def _prepare_circuit(self, circuit: "Circuit", keys: list[int]) -> tuple[NDArray, list[int], NDArray]:
-        """Prepare the circuit for execution by constructing the necessary state and transformation matrices.
-        
-        Args:
-            circuit (Circuit): quantum circuit to apply.
-            keys (list[int]): list of keys for quantum states to apply circuit to.
-        
-        Returns:
-            tuple: tuple containing the new state, all keys, and the circuit matrix.
-                   Note: the returned circuit matrix contains any necessary swaps to align qubits of new state
-        """
-        old_states = []
-        all_keys = []
-
-        # go through keys and get all unique qstate objects
-        for key in keys:
-            qstate = self.states[key]
-            if qstate.keys[0] not in all_keys:
-                old_states.append(qstate.state)
-                all_keys += qstate.keys
-
-        # construct compound state; order qubits
-        new_state = [1]
-        for state in old_states:
-            new_state = kron(new_state, state)
-
-        # get circuit matrix; expand if necessary
-        circ_mat = circuit.get_unitary_matrix()
-        if circuit.size < len(all_keys):
-            # pad size of circuit matrix if necessary
-            diff = len(all_keys) - circuit.size
-            circ_mat = kron(circ_mat, identity(2 ** diff))
-
-        # apply any necessary swaps
-        if not all([all_keys.index(key) == i for i, key in enumerate(keys)]):
-            all_keys, swap_mat = self._swap_qubits(all_keys, keys)
-            circ_mat = circ_mat @ swap_mat
-
-        return new_state, all_keys, circ_mat
-
-    @staticmethod
-    def _swap_qubits(all_keys: list[int], keys: list[int]) -> tuple[list[int], NDArray]:
-        """Swap qubits in the circuit.
-        
-        Args:
-            all_keys (list[int]): The list of all qubit keys.
-            keys (list[int]): The list of qubit keys to swap.
-        
-        Returns:
-            tuple: updated list of all keys and the swap matrix.
-        """
-        swap_circuit = QubitCircuit(N=len(all_keys))
-        for i, key in enumerate(keys):
-            j = all_keys.index(key)
-            if j != i:
-                gate = Gate("SWAP", targets=[i, j])
-                swap_circuit.add_gate(gate)
-                all_keys[i], all_keys[j] = all_keys[j], all_keys[i]
-        swap_mat = gate_sequence_product(swap_circuit.propagators()).full()
-        return all_keys, swap_mat
-
-    @abstractmethod
-    def set(self, keys: list[int], amplitudes: Any) -> None:
-        """Method to set quantum state at a given key(s).
-
-        Args:
-            keys (list[int]): key(s) of state(s) to change.
-            amplitudes (any): Amplitudes to set state to, type determined by type of subclass.
-        """
-
-        pass
-
-    def remove(self, key: int) -> None:
-        """Method to remove state stored at key.
-        
-        Args:
-            key (int): The key of the state to remove.
-        """
-        del self.states[key]
-
-    def set_states(self, states: dict):
-        """Set multiple quantum states.
-
-        Args:
-            states (dict): A dictionary mapping keys to their corresponding quantum states.
-        """
-        self.states = states
-
-
-@QuantumManager.register(KET_VECTOR_FORMALISM)
-class QuantumManagerKet(QuantumManager):
-    """Class to track and manage quantum states with the ket vector formalism."""
-
-    def __init__(self, **kwargs):
-        super().__init__()
-
-    def new(self, state=[complex(1), complex(0)]) -> int:
-        """Method to create a new ket state.
-
-        Args:
-            state (list[complex]): amplitudes of new state.
-
-        Returns:
-            int: the key of the new state.
-        """
-        key = self._least_available
-        self._least_available += 1
-        self.states[key] = KetState(state, [key])
-        return key
-
-    def run_circuit(self, circuit: "Circuit", keys: list[int], meas_samp=None) -> dict[int, int]:
-        """Method to run a circuit on a given list of keys.
-        
-        Args:
-            circuit (Circuit): quantum circuit to apply.
-            keys (list[int]): list of keys to apply circuit to.
-            meas_samp (float): random sample used for measurement result.
-        
-        Returns:
-            If measurement, dict[int, int]: dictionary mapping qstate keys to measurement results.
-            If non-measurement, dict: empty dictionary.
-        """
-        super().run_circuit(circuit, keys, meas_samp)
-        new_state, all_keys, circ_mat = self._prepare_circuit(circuit, keys)
-
-        new_state = circ_mat @ new_state
-
-        if len(circuit.measured_qubits) == 0:
-            # set state, return no measurement result
-            new_ket = KetState(new_state, all_keys)
-            for key in all_keys:
-                self.states[key] = new_ket
-            return {}
-        else:
-            # measure state (state reassignment done in _measure method)
-            keys = [all_keys[i] for i in circuit.measured_qubits]
-            return self._measure(new_state, keys, all_keys, meas_samp)
-
-    def set(self, keys: list[int], amplitudes: list[complex]) -> None:
-        """Set the quantum state for the given keys.
-
-        Args:
-            keys (list[int]): list of keys of the quantum state.
-            amplitudes (list[complex]): amplitudes to set the state to.
-        """
-        super().set(keys, amplitudes)
-        new_state = KetState(amplitudes, keys)
-        for key in keys:
-            self.states[key] = new_state
-
-    def get_ascending_keys(self, key: int) -> KetState:
-        """Method to get quantum state stored at an index.
-           Reorders qubits (in-place) in ascending order of keys before returning.
-
-        Args:
-            key (int): key for quantum state.
-
-        Returns:
-            KetState: quantum state at supplied key.
-        """
-        state = super().get(key)
-        self.reorder_qubits_ascending_keys(state)
-        return state
-
-    def reorder_qubits_ascending_keys(self, state: KetState) -> None:
-        """Update the quantum state (in-place) to match the ascending order of keys.
-           Meanwhile, the reordered state is also set in the quantum manager.
-        
-        Args:
-            state (KetState): The quantum state to reorder.
-        """
-        target_all_keys = sorted(state.keys)
-        if state.keys != target_all_keys:
-            _, swap_matrix = self._swap_qubits(state.keys, target_all_keys)
-            reordered_state = swap_matrix @ state.state
-            state.state = reordered_state
-            self.set(target_all_keys, reordered_state)
-
-    def set_to_zero(self, key: int) -> None:
-        """Set the qubit at the given key to the |0> state.
-
-        Args:
-            key (int): key of the qubit to set to |0>.
-        """
-        self.set([key], [complex(1), complex(0)])
-
-    def set_to_one(self, key: int) -> None:
-        """Set the qubit at the given key to the |1> state.
-
-        Args:
-            key (int): key of the qubit to set to |1>.
-        """
-        self.set([key], [complex(0), complex(1)])
-
-    def _measure(self, state: list[complex], keys: list[int], all_keys: list[int], meas_samp: float) -> dict[int, int]:
-        """Method to measure qubits at given keys.
-
-        SHOULD NOT be called individually; only from circuit method (unless for unit testing purposes).
-        Modifies quantum state of all qubits given by all_keys.
-
-        Args:
-            state (list[complex]): state to measure.
-            keys (list[int]): list of keys to measure.
-            all_keys (list[int]): list of all keys corresponding to state.
-            meas_samp (float): random number between 0 and 1 used for measurement.
-
-        Returns:
-            dict[int, int]: mapping of measured keys to measurement results.
-        """
-        if len(keys) == 1:
-            if len(all_keys) == 1:
-                prob_0 = measure_state_with_cache_ket(tuple(state))
-                if meas_samp < prob_0:
-                    result = 0
-                else:
-                    result = 1
-            else:
-                key = keys[0]
-                num_states = len(all_keys)
-                state_index = all_keys.index(key)
-                state_0, state_1, prob_0 = measure_entangled_state_with_cache_ket(tuple(state), state_index, num_states)
-                if meas_samp < prob_0:
-                    new_state = array(state_0, dtype=complex)
-                    result = 0
-                else:
-                    new_state = array(state_1, dtype=complex)
-                    result = 1
-
-            all_keys.remove(keys[0])
-
-        else:
-            # swap states into correct position
-            if not all([all_keys.index(key) == i for i, key in enumerate(keys)]):
-                all_keys, swap_mat = self._swap_qubits(all_keys, keys)
-                state = swap_mat @ state
-
-            # calculate meas probabilities and projected states
-            len_diff = len(all_keys) - len(keys)
-            new_states, probabilities = measure_multiple_with_cache_ket(tuple(state), len(keys), len_diff)
-
-            # choose result, set as new state
-            for i in range(int(2 ** len(keys))):
-                if meas_samp < sum(probabilities[:i + 1]):
-                    result = i
-                    new_state = new_states[i]
-                    break
-
-            for key in keys:
-                all_keys.remove(key)
-
-        result_states = [array([1, 0]), array([0, 1])]
-        result_digits = [int(x) for x in bin(result)[2:]]
-        while len(result_digits) < len(keys):
-            result_digits.insert(0, 0)
-
-        for res, key in zip(result_digits, keys):
-            # set to state measured
-            new_state_obj = KetState(result_states[res], [key])
-            self.states[key] = new_state_obj
-
-        if len(all_keys) > 0:
-            new_state_obj = KetState(new_state, all_keys)
-            for key in all_keys:
-                self.states[key] = new_state_obj
-
-        return dict(zip(keys, result_digits))
-
-
-@QuantumManager.register(DENSITY_MATRIX_FORMALISM)
-class QuantumManagerDensity(QuantumManager):
-    """Class to track and manage states with the density matrix formalism."""
-
-    def __init__(self, **kwargs):
-        super().__init__()
-
-    def new(self, state=([complex(1), complex(0)], [complex(0), complex(0)])) -> int:
-        """Method to create a new density matrix state.
-        
-        Args:
-            state (list[list[complex]]): density matrix state.
-
-        Returns:
-            int: key of the new state.
-        """
-        key = self._least_available
-        self._least_available += 1
-        self.states[key] = DensityState(state, [key])
-        return key
-
-    def run_circuit(self, circuit: "Circuit", keys: list[int], meas_samp=None) -> dict[int, int]:
-        """Method to run a circuit on a given list of keys.
-        
-        Args:
-            circuit (Circuit): quantum circuit to apply.
-            keys (list[int]): list of keys to apply circuit to.
-            meas_samp (float): random number between 0 and 1 used for measurement.
-
-        Returns:
-            If measurement, dict[int, int]: dictionary mapping qstate keys to measurement results.
-            If non-measurement, dict: empty dictionary.
-        """
-        super().run_circuit(circuit, keys, meas_samp)
-        new_state, all_keys, circ_mat = super()._prepare_circuit(circuit, keys)
-
-        new_state = circ_mat @ new_state @ circ_mat.conj().T
-
-        if len(circuit.measured_qubits) == 0:
-            # set state, return no measurement result
-            new_state_obj = DensityState(new_state, all_keys)
-            for key in all_keys:
-                self.states[key] = new_state_obj
-            return {}
-        else:
-            # measure state (state reassignment done in _measure method)
-            keys = [all_keys[i] for i in circuit.measured_qubits]
-            return self._measure(new_state, keys, all_keys, meas_samp)
-
-    def set(self, keys: list[int], state: list[list[complex]]) -> None:
-        """Method to set the quantum state at the given keys.
-
-        The `state` argument should be passed as list[list[complex]], where each internal list is a row.
-        However, the `state` may also be given as a one-dimensional pure state.
-        If the list is one-dimensional, will be converted to matrix with the outer product operation.
-
-        Args:
-            keys (list[int]): list of quantum manager keys to modify.
-            state: quantum state to set input keys to.
-        """
-        super().set(keys, state)
-        new_state = DensityState(state, keys)
-        for key in keys:
-            self.states[key] = new_state
-
-    def set_to_zero(self, key: int):
-        """Set the qubit at the given key to the |0><0| state.
-        
-        Args:
-            key (int): key of the qubit to set to |0><0|.
-        """
-        self.set([key], [[complex(1), complex(0)], [complex(0), complex(0)]])
-
-    def set_to_one(self, key: int):
-        """Set the qubit at the given key to the |1><1| state.
-        
-        Args:
-            key (int): key of the qubit to set to |1><1|.
-        """
-        self.set([key], [[complex(0), complex(0)], [complex(0), complex(1)]])
-
-    def get_ascending_keys(self, key: int) -> DensityState:
-        """Method to get quantum state stored at an index.
-           Reorders qubits (in-place) in ascending order of keys before returning.
-
-        Args:
-            key (int): key for quantum state.
-
-        Returns:
-            DensityState: quantum state at supplied key.
-        """
-        state = super().get(key)
-        self.reorder_qubits_ascending_keys(state)
-        return state
-
-    def reorder_qubits_ascending_keys(self, state: DensityState) -> None:
-        """Update the quantum state (in-place) to match the ascending order of keys.
-           Meanwhile, the reordered state is also set in the quantum manager.
-        
-        Args:
-            state (DensityState): The quantum state to reorder.
-        """
-        target_all_keys = sorted(state.keys)
-        if state.keys != target_all_keys:
-            _, swap_matrix = self._swap_qubits(state.keys, target_all_keys)
-            reordered_state = swap_matrix @ state.state @ swap_matrix.conj().T
-            state.state = reordered_state
-            self.set(target_all_keys, reordered_state)
-
-    def _measure(self, state: list[list[complex]], keys: list[int], all_keys: list[int], meas_samp: float) -> dict[int, int]:
-        """Method to measure qubits at given keys.
-
-        SHOULD NOT be called individually; only from circuit method (unless for unit testing purposes).
-        Modifies quantum state of all qubits given by all_keys.
-
-        Args:
-            state (list[complex]): state to measure.
-            keys (list[int]): list of keys to measure.
-            all_keys (list[int]): list of all keys corresponding to state.
-            meas_samp (float): random number between 0 and 1 used for measurement.
-
-        Returns:
-            dict[int, int]: mapping of measured keys to measurement results.
-        """
-
-        if len(keys) == 1:
-            if len(all_keys) == 1:
-                prob_0 = measure_state_with_cache_density(tuple(map(tuple, state)))
-                if meas_samp < prob_0:
-                    result = 0
-                    new_state = [[1, 0], [0, 0]]
-                else:
-                    result = 1
-                    new_state = [[0, 0], [0, 1]]
-
-            else:
-                key = keys[0]
-                num_states = len(all_keys)
-                state_index = all_keys.index(key)
-                state_0, state_1, prob_0 = measure_entangled_state_with_cache_density(tuple(map(tuple, state)), 
-                                                                                      state_index, num_states)
-                if meas_samp < prob_0:
-                    new_state = array(state_0, dtype=complex)
-                    result = 0
-                else:
-                    new_state = array(state_1, dtype=complex)
-                    result = 1
-
-        else:
-            # swap states into correct position
-            if not all([all_keys.index(key) == i for i, key in enumerate(keys)]):
-                all_keys, swap_mat = self._swap_qubits(all_keys, keys)
-                state = swap_mat @ state @ swap_mat.conj().T
-
-            # calculate meas probabilities and projected states
-            len_diff = len(all_keys) - len(keys)
-            state_to_measure = tuple(map(tuple, state))
-            new_states, probabilities = measure_multiple_with_cache_density(state_to_measure, len(keys), len_diff)
-
-            # choose result, set as new state
-            for i in range(int(2 ** len(keys))):
-                if meas_samp < sum(probabilities[:i + 1]):
-                    result = i
-                    new_state = new_states[i]
-                    break
-
-        result_digits = [int(x) for x in bin(result)[2:]]
-        while len(result_digits) < len(keys):
-            result_digits.insert(0, 0)
-
-        new_state_obj = DensityState(new_state, all_keys)
-        for key in all_keys:
-            self.states[key] = new_state_obj
-
-        return dict(zip(keys, result_digits))
-
-
-@QuantumManager.register(FOCK_DENSITY_MATRIX_FORMALISM)
-class QuantumManagerDensityFock(QuantumManager):
-    """Class to track and manage Fock states with the density matrix formalism."""
-
-    def __init__(self, truncation: int = 1, **kwargs):
-        # default truncation is 1 for 2-d Fock space.
-        super().__init__(truncation=truncation)
-
-    def new(self, state=None) -> int:
-        """Method to create a new state with key
-
-        Args:
-            state (str | list[complex] | list[list[complex]]): amplitudes of new state.
-                Default value is 'gnd': create zero-excitation state with current truncation.
-                Other inputs are passed to the constructor of `DensityState`.
-        """
-
-        key = self._least_available
-        self._least_available += 1
-        if state is None:
-            gnd = [1] + [0] * self.truncation
-            self.states[key] = DensityState(gnd, [key], truncation=self.truncation)
-        else:
-            self.states[key] = DensityState(state, [key], truncation=self.truncation)
-
-        return key
-
-    def run_circuit(self, circuit: "Circuit", keys: list[int], meas_samp=None) -> dict[int, int]:
-        """Currently the Fock states do not support quantum circuits.
-        This method is only to implement abstract method of parent class and SHOULD NOT be called after instantiation.
-        """
-        raise Exception("run_circuit method of class QuantumManagerDensityFock called")
-
-    def _generate_swap_operator(self, num_systems: int, i: int, j: int):
-        """Helper function to generate swapping unitary.
-
-        Args:
-            num_systems (int): number of subsystems in state
-            i (int): index of first subsystem to swap
-            j (int): index of second subsystem to swap
-
-        Returns:
-            Array[int]: unitary swapping operator
-        """
-
-        size = self.dim ** num_systems
-        swap_unitary = zeros((size, size))
-
-        for old_index in range(size):
-            old_str = base_repr(old_index, self.dim)
-            old_str = old_str.zfill(num_systems)
-            new_str = ''.join((old_str[:i], old_str[j], old_str[i + 1:j], old_str[i], old_str[j + 1:]))
-            new_index = int(new_str, base=self.dim)
-            swap_unitary[new_index, old_index] = 1
-
-        return swap_unitary
-
-    def _prepare_state(self, keys: list[int]):
-        """Function to prepare states at given keys for operator application.
-
-        Will take composite quantum state and swap subsystems to correspond with listed keys.
-        Should not be called directly, but from method to apply operator or measure state.
-
-        Args:
-            keys (list[int]): keys for states to apply operator to.
-
-        Returns:
-            tuple(list[list[complex]], list[int]): tuple containing:
-                1. new state to apply operator to, with keys swapped to be consecutive.
-                2. list of keys corresponding to new state.
-        """
-
-        old_states = []
-        all_keys = []
-
-        # go through keys and get all unique qstate objects
-        for key in keys:
-            qstate = self.states[key]
-            if qstate.keys[0] not in all_keys:
-                old_states.append(qstate.state)
-                all_keys += qstate.keys
-
-        # construct compound state
-        new_state = [1]
-        for state in old_states:
-            new_state = kron(new_state, state)
-
-        # apply any necessary swaps to order keys
-        if len(keys) > 1:
-
-            # generate desired key order
-            start_idx = all_keys.index(keys[0])
-            if start_idx + len(keys) > len(all_keys):
-                start_idx = len(all_keys) - len(keys)
-
-            for i, key in enumerate(keys):
-                i = i + start_idx
-                j = all_keys.index(key)
-                if j != i:
-                    swap_unitary = self._generate_swap_operator(len(all_keys), i, j)
-                    new_state = swap_unitary @ new_state @ swap_unitary.T
-                    all_keys[i], all_keys[j] = all_keys[j], all_keys[i]
-
-        return new_state, all_keys
-
-    def _prepare_operator(self, all_keys: list[int], keys: list[int], operator) -> NDArray:
-        # pad operator with identity
-        left_dim = self.dim ** all_keys.index(keys[0])
-        right_dim = self.dim ** (len(all_keys) - all_keys.index(keys[-1]) - 1)
-        prepared_operator = operator
-
-        if left_dim > 0:
-            prepared_operator = kron(identity(left_dim), prepared_operator)
-        if right_dim > 0:
-            prepared_operator = kron(prepared_operator, identity(right_dim))
-
-        return prepared_operator
-
-    def apply_operator(self, operator: NDArray, keys: list[int]):
-        prepared_state, all_keys = self._prepare_state(keys)
-        prepared_operator = self._prepare_operator(all_keys, keys, operator)
-        new_state = prepared_operator @ prepared_state @ prepared_operator.conj().T
-        self.set(all_keys, new_state)
-
-    def set(self, keys: list[int], state: list[list[complex]]) -> None:
-        """Method to set the quantum state at the given keys.
-
-        The `state` argument should be passed as list[list[complex]], where each internal list is a row.
-        However, the `state` may also be given as a one-dimensional pure state.
-        If the list is one-dimensional, will be converted to matrix with the outer product operation.
-
-        Args:
-            keys (list[int]): list of quantum manager keys to modify.
-            state: quantum state to set input keys to.
-        """
-
-        super().set(keys, state)
-        new_state = DensityState(state, keys, truncation=self.truncation)
-        for key in keys:
-            self.states[key] = new_state
-
-    def set_to_zero(self, key: int):
-        """set the state to ground (zero) state."""
-        gnd = [1] + [0] * self.truncation
-        self.set([key], gnd)
-
-    def build_ladder(self):
-        """Generate matrix of creation and annihilation (ladder) operators on truncated Hilbert space."""
-        truncation = self.truncation
-        data = array([sqrt(i + 1) for i in range(truncation)])  # elements in create/annihilation operator matrix
-        row = array([i + 1 for i in range(truncation)])
-        col = array([i for i in range(truncation)])
-        create = csr_matrix((data, (row, col)), shape=(truncation + 1, truncation + 1)).toarray()
-        destroy = create.conj().T
-
-        return create, destroy
-
-    def measure(self, keys: list[int], povms: list[NDArray], meas_samp: float) -> int:
-        """Method to measure subsystems at given keys in POVM formalism.
-
-        Serves as wrapper for private `_measure` method, performing quantum manager specific operations.
-
-        Args:
-            keys (list[int]): list of keys to measure.
-            povms: (list[array]): list of POVM operators to use for measurement.
-            meas_samp (float): random measurement sample to use for computing resultant state.
-
-        Returns:
-            int: measurement as index of matching POVM in supplied tuple.
-        """
-
-        new_state, all_keys = self._prepare_state(keys)
-        return self._measure(new_state, keys, all_keys, povms, meas_samp)
-
-    def _measure(self, state: list[list[complex]], keys: list[int],
-                 all_keys: list[int], povms: list[NDArray], meas_samp: float) -> int:
-        """Method to measure subsystems at given keys in POVM formalism.
-
-        Modifies quantum state of all qubits given by all_keys, post-measurement operator determined
-        by measurement operators which are chosen as square root of POVM operators.
-
-        Args:
-            state (list[list[complex]]): state to measure.
-            keys (list[int]): list of keys to measure.
-            all_keys (list[int]): list of all keys corresponding to state.
-            povms: (list[NDArray]): list of POVM operators to use for measurement.
-            meas_samp (float): random measurement sample to use for computing resultant state.
-
-        Returns:
-            int: measurement as index of matching POVM in supplied tuple.
-        """
-
-        state_tuple = tuple(map(tuple, state))
-        povm_tuple = tuple([tuple(map(tuple, povm)) for povm in povms])
-        new_state = None
-        result = 0
-
-        # calculate meas probabilities and projected states
-        if len(keys) == 1:
-            if len(all_keys) == 1:
-                states, probs = measure_state_with_cache_fock_density(state_tuple, povm_tuple)
-
-            else:
-                key = keys[0]
-                num_states = len(all_keys)
-                state_index = all_keys.index(key)
-                states, probs = measure_entangled_state_with_cache_fock_density(state_tuple, state_index, num_states, 
-                                                                                povm_tuple, self.truncation)
-
-        else:
-            indices = tuple([all_keys.index(key) for key in keys])
-            states, probs = measure_multiple_with_cache_fock_density(state_tuple, indices, len(all_keys), 
-                                                                     povm_tuple, self.truncation)
-
-        # calculate result based on measurement sample.
-        prob_sum = cumsum(probs)
-        for i, (output_state, p) in enumerate(zip(states, prob_sum)):
-            if meas_samp < p:
-                new_state = output_state
-                result = i
-                break
-
-        """
-        # for potential future work
-        result_digits = [int(x) for x in base_repr(result, base=self.dim)[2:]]
-        while len(result_digits) < len(keys):
-            result_digits.insert(0, 0)
-
-        # assign measured states
-        for key, result in zip(keys, result_digits):
-            state = [0] * self.dim
-            state[result] = 1
-            self.set([key], state)
-        """
-
-        for key in keys:
-            self.states[key] = None  # clear the stored state at key (particle destructively measured)
-
-        # assign remaining state
-        if len(keys) < len(all_keys):
-            indices = tuple([all_keys.index(key) for key in keys])
-            new_state_tuple = tuple(map(tuple, new_state))
-            remaining_state = density_partial_trace(new_state_tuple, indices, len(all_keys), self.truncation)
-            remaining_keys = [key for key in all_keys if key not in keys]
-            self.set(remaining_keys, remaining_state)
-
-        return result
-
-    def _build_loss_kraus_operators(self, loss_rate: float, all_keys: list[int], key: int) -> list[array]:
-        """Method to build Kraus operators of a generalized amplitude damping channel.
-
-        This represents the effect of photon loss.
-
-        Args:
-            loss_rate (float): loss rate for the quantum channel.
-            all_keys (list[int]): list of all keys in affected state.
-            key (int): key for subsystem experiencing loss.
-
-        Returns:
-            list[array]: list of generated Kraus operators.
-        """
-
-        assert 0 <= loss_rate <= 1
-        kraus_ops = []
-
-        for k in range(self.dim):
-            total_kraus_op = zeros((self.dim ** len(all_keys), self.dim ** len(all_keys)))
-
-            for n in range(k, self.dim):
-                coeff = sqrt(binom(n, k)) * sqrt(((1 - loss_rate) ** (n - k)) * (loss_rate ** k))
-                single_op = zeros((self.dim, self.dim))
-                single_op[n - k, n] = 1
-                total_op = self._prepare_operator(all_keys, [key], single_op)
-                total_kraus_op += coeff * total_op
-
-            kraus_ops.append(total_kraus_op)
-
-        return kraus_ops
-
-    def add_loss(self, key, loss_rate):
-        """Method to apply generalized amplitude damping channel on a *single* subspace corresponding to `key`.
-
-        Args:
-            key (int): key for the subspace experiencing loss.
-            loss_rate (float): loss rate for the quantum channel.
-        """
-
-        prepared_state, all_keys = self._prepare_state([key])
-        kraus_ops = self._build_loss_kraus_operators(loss_rate, all_keys, key)
-        output_state = zeros(prepared_state.shape, dtype=complex)
-
-        for kraus_op in kraus_ops:
-            output_state += kraus_op @ prepared_state @ kraus_op.conj().T
-
-        self.set(all_keys, output_state)
-
-
-@QuantumManager.register(BELL_DIAGONAL_STATE_FORMALISM)
-class QuantumManagerBellDiagonal(QuantumManager):
-    """Class to track and manage quantum states with the bell diagonal formalism.
-
-    To be aligned with analytical formulae, we have assumed that successfully generated EPR pair is in Phi+ form.
-    And note that the 4 BDS elements are in I, Z, X, Y order.
-
-    * BDS is only used for entanglement distribution (generation, swapping, purification), assuming underlying errors being purely Pauli.
-    * All manipulation results can be tracked analytically, without explicit quantum gates / channels / measurements.
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__()
-
-    def new(self, state=None) -> int:
-        """Generates new quantum state key for quantum manager.
-
-        NOTE: since this generates only one state, there will be no corresponding entangled state stored.
-        The Bell diagonal state formalism assumes entangled states;
-        thus, attempting to call `get` will return an exception until entangled.
-        The purpose of this function is thus mainly to avoid state key collisions.
-
-        Args:
-            state (Any): to conform to type definition (does nothing).
-
-        Returns:
-            int: quantum state key corresponding to state.
-        """
-        key = self._least_available
-        self._least_available += 1
-        return key
-
-    def get(self, key: int):
-        if key not in self.states:
-            raise Exception("Attempt to get Bell diagonal state before entanglement.")
-
-        return super().get(key)
-
-    def set(self, keys: list[int], diag_elems: list[float]) -> None:
-        super().set(keys, diag_elems)
-        # assert len(keys) == 2, "Bell diagonal states must have 2 keys."
-        if len(keys) != 2:
-            # raise Warning("bell diagonal quantum manager received invalid set request")  # optional
-            for key in keys:
-                if key in self.states:
-                    self.states.pop(key)
-            return
-        new_state = BellDiagonalState(diag_elems, keys)
-        for key in keys:
-            self.states[key] = new_state
-
-    def set_to_noiseless(self, keys: list[int]):
-        self.set(keys, [float(1), float(0), float(0), float(0)])
-
-    def run_circuit(self, *args, **kwargs):
-        pass
